+++ obj/src/clients/CloudFunctionClient.d.ts
@@ -26,7 +26,10 @@ import { GcpConnectionResolver } from '../connect/GcpConnectionResolver';
  *      - region:        is the region where your function is deployed
  *      - function:      is the name of the HTTP function you deployed
  *      - org_id:        organization name
- *
+ * - options:
+ *      - retries:               number of retries (default: 3)
+ *      - connect_timeout:       connection timeout in milliseconds (default: 10 sec)
+ *      - timeout:               invocation timeout in milliseconds (default: 10 sec)
  * - credentials:
  *     - account: the service account name
  *     - auth_token:    Google-generated ID token or null if using custom auth (IAM)
@@ -69,6 +72,7 @@ import { GcpConnectionResolver } from '../connect/GcpConnectionResolver';
  *     const result = await client.getData("123", "1");
  */
 export declare abstract class CloudFunctionClient implements IOpenable, IConfigurable, IReferenceable {
+    private static readonly _defaultConfig;
     /**
      * The HTTP client.
      */