+++ obj/test/services/DummyCloudFunctionService.d.ts
@@ -3,8 +3,11 @@ import { CloudFunctionService } from '../../src/services/CloudFunctionService';
 export declare class DummyCloudFunctionService extends CloudFunctionService {
     private _controller;
     private _headers;
+    private numberOfCalls;
     constructor();
     setReferences(references: IReferences): void;
+    private incrementNumberOfCalls;
+    private getNumberOfCalls;
     private getPageByFilter;
     private getOneById;
     private create;