+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200">
 <br/> Google Cloud Platform specific components for Node.js / ES2017 Changelog
 
+## <a name="1.0.5"></a> 1.0.5 (2022-09-07)
+
+### Features
+* Added interceptors by command name and regex
+* Added more config parameters for clients
+
+### Bug Fixes
+* Fixed error responses
+
 ## <a name="1.0.1-1.0.4"></a> 1.0.1-1.0.4 (2022-03-09)
 
 ### Bug Fixes